+++ packages/common/test/index.test.js
@@ -31,6 +31,8 @@ import {
   toArray,
   validate,
   assert as assertCommon,
+  log,
+  debug,
 } from '../src/Adaptor';
 import { startOfToday } from 'date-fns';
 
@@ -1108,5 +1110,104 @@ describe('assert', () => {
     }
 
     expect(error.message).to.eql(`assertion statement failed with false`);
+  });
+});
+
+describe('log', () => {
+  const input = { x: 'a', y: { z: 'b' }, u: null };
+  let originalLog;
+  let consoleOutput = [];
+  beforeEach(() => {
+    originalLog = console.log;
+    console.log = (...args) => consoleOutput.push(...args);
+  });
+
+  it('should log a message', () => {
+    const result = log('test')(input);
+    expect(result).to.eq(input);
+    expect(consoleOutput[0]).to.eq('test');
+  });
+
+  it('should log multiple messages', () => {
+    log('test', 'test2')(input);
+    expect(consoleOutput[0]).to.eq('test');
+    expect(consoleOutput[1]).to.eq('test2');
+  });
+  it('should log multiple messages with state', () => {
+    log(
+      'test',
+      state => state.x,
+      state => `test ${state.y.z}`
+    )(input);
+    expect(consoleOutput[0]).to.eq('test');
+    expect(consoleOutput[1]).to.eq('a');
+    expect(consoleOutput[2]).to.eq('test b');
+  });
+  it('should log undefined if key is not found', () => {
+    log(state => state.a)({});
+    expect(consoleOutput[0]).to.eq(undefined);
+  });
+  it('should log the value of a state reference', () => {
+    log(state => state.x)(input);
+    log(state => state.u)(input);
+    expect(consoleOutput[0]).to.eq('a');
+    expect(consoleOutput[1]).to.eq(null);
+  });
+  afterEach(() => {
+    consoleOutput = [];
+    console.log = originalLog;
+  });
+});
+
+describe('debug', () => {
+  const input = { x: 'a', y: { z: 'b' }, u: null };
+  let originalDebug;
+  let consoleOutput = [];
+
+  beforeEach(() => {
+    originalDebug = console.debug;
+    console.debug = (...args) => consoleOutput.push(...args);
+  });
+
+  it('should debug a message', () => {
+    const result = debug('test')(input);
+    expect(result).to.eq(input);
+    expect(consoleOutput[0]).to.eq('test');
+  });
+
+  it('should debug multiple messages', () => {
+    debug('test', 'test2')(input);
+    expect(consoleOutput[0]).to.eq('test');
+    expect(consoleOutput[1]).to.eq('test2');
+  });
+
+  it('should debug multiple messages with state', () => {
+    debug(
+      'test',
+      state => state.x,
+      state => `test ${state.y.z}`
+    )(input);
+
+    expect(consoleOutput[0]).to.eq('test');
+    expect(consoleOutput[1]).to.eq('a');
+    expect(consoleOutput[2]).to.eq('test b');
+  });
+
+  it('should debug undefined if key is not found', () => {
+    debug(state => state.a)({});
+    expect(consoleOutput[0]).to.eq(undefined);
+  });
+
+  it('should debug the value of a state reference', () => {
+    debug(state => state.x)(input);
+    debug(state => state.u)(input);
+
+    expect(consoleOutput[0]).to.eq('a');
+    expect(consoleOutput[1]).to.eq(null);
+  });
+
+  afterEach(() => {
+    consoleOutput = [];
+    console.debug = originalDebug;
   });
 });