+++ packages/common/test/util/http.test.js
@@ -9,7 +9,9 @@ import {
   del,
   parseUrl,
   ERROR_URL_MISMATCH,
+  logResponse,
 } from '../../src/util/http.js';
+import { encode } from '../../src/util/base64.js';
 
 const client = enableMockClient('https://www.example.com');
 
@@ -118,6 +120,43 @@ describe('parseUrl', () => {
     } catch (e) {
       expect(e.message).to.eql(ERROR_URL_MISMATCH);
     }
+  });
+});
+
+describe('logResponse', () => {
+  it('should include query parameters in the url', async () => {
+    client
+      .intercept({
+        path: '/api',
+        method: 'GET',
+        query: {
+          name: 'homelander',
+        },
+      })
+      .reply(200, {});
+
+    const response = await request('GET', 'https://www.example.com/api', {
+      query: {
+        name: 'homelander',
+      },
+    });
+    response.duration = 2;
+    let originalLog;
+    let consoleOutput = [];
+    
+    // Setup: Override console.log
+    originalLog = console.log;
+    console.log = (...args) => consoleOutput.push(args);
+
+    logResponse(response, {
+      name: 'homelander',
+    });
+    expect(consoleOutput?.[0]).to.deep.eql([
+      'GET https://www.example.com/api?name=homelander - 200 in 2ms',
+    ]);
+
+    // Teardown: Restore console.log
+    console.log = originalLog;
   });
 });
 
@@ -671,7 +710,7 @@ describe('helpers', () => {
   });
 
   describe('parseAs', () => {
-    it('should auto parse as json', async () => {
+    it('should auto parse as json if content-type is application/json', async () => {
       client
         .intercept({
           path: '/api',
@@ -694,7 +733,7 @@ describe('helpers', () => {
       expect(result.body).to.eql({ name: 'mutchi' });
     });
 
-    it('should auto parse as text by default', async () => {
+    it('should auto parse as text by default (no content type)', async () => {
       client
         .intercept({
           path: '/api',
@@ -809,6 +848,24 @@ describe('helpers', () => {
       expect(body).to.eql({
         id: '2',
       });
+    });
+
+    it('should force as base64', async () => {
+      const binaryData = Buffer.from('This is binary content', 'utf8');
+      client
+        .intercept({
+          path: '/api',
+          method: 'GET',
+        })
+        .reply(200, binaryData);
+
+      const result = await request('GET', 'https://www.example.com/api', {
+        parseAs: 'base64',
+      });
+
+      const base64Encoded = encode(binaryData, { parseJson: false });
+
+      expect(result.body).to.eql(base64Encoded);
     });
   });
 });