+++ packages/http/src/Adaptor.js
@@ -21,6 +21,7 @@ import { request as sendRequest, xmlParser } from './util';
  * @property data - the parsed response body
  * @property response - the response from the HTTP server, including headers, statusCode, body, etc
  * @property references - an array of all previous data objects used in the Job
+ * @private
  **/
 
 /**
@@ -51,7 +52,14 @@ export function execute(...operations) {
  * @example <caption>Make a GET request</caption>
  * request('GET', '/patient', {
  *   query: { foo: 'bar', a: 1 },
- *   headers: { 'content-type': 'application/json' },
+ * });
+ * @example <caption>Make a POST request with a body</caption>
+ * request('POST', '/todos', {
+ *   body:{
+ *     "userId": 1,
+ *     "title": "delectus aut autem",
+ *     "completed": false
+ *   },
  * });
  * @function
  * @param {string} method - The HTTP method to use.
@@ -70,7 +78,6 @@ export function request(method, path, options) {
  * @example <caption>GET request with query parameters and custom headers</caption>
  * get('/patient', {
  *   query: { foo: 'bar', a: 1 },
- *   headers: { 'content-type': 'application/json' },
  * });
  * @function
  * @param {string} path - Path to resource. Can be an absolute URL if baseURL is NOT set on `state.configuration`.