+++ packages/hubtel/src/Adaptor.js
@@ -58,6 +58,7 @@ export function request(method, path, body, options = {}) {
  * @property data - the parsed response body
  * @property response - the response from the HTTP server, including headers, statusCode, body, etc
  * @property references - an array of all previous data objects used in the Job
+ * @private
  **/
 
 /**