+++ packages/odk/src/Adaptor.js
@@ -11,6 +11,7 @@ import * as util from './Utils';
  * @property data - the parsed response body
  * @property response - the response from the ODK HTTP server (with the body removed)
  * @property references - an array of all the previous data values
+ * @private
  **/
 
 /**