+++ tools/generate/template/src/Adaptor.js
@@ -87,15 +87,20 @@ export function request(method, path, body, options = {}) {
 }
 
 export {
+  combine,
+  cursor,
   dataPath,
   dataValue,
   dateFns,
-  cursor,
   each,
   field,
   fields,
   fn,
+  fnIf,
+  group,
   lastReferenceValue,
   merge,
+  scrubEmojis,
   sourceValue,
+  util,
 } from '@openfn/language-common';