+++ packages/divoc/CHANGELOG.md
@@ -1,6 +1,30 @@
 # @openfn/language-divoc
 
-## 0.1.0
+## 0.1.3 - 22 April 2025
+
+### Patch Changes
+
+- 99e4b48: - Better handling of HTML content in http requests
+  - When logging HTTP requests, include query parameters
+- Updated dependencies \[99e4b48]
+- Updated dependencies \[13bf08f]
+  - @openfn/language-common@2.4.0
+
+## 0.1.2 - 16 April 2025
+
+### Patch Changes
+
+- Updated dependencies \[b089c56]
+  - @openfn/language-common@2.3.3
+
+## 0.1.1 - 11 April 2025
+
+### Patch Changes
+
+- Updated dependencies \[d7105c0]
+  - @openfn/language-common@2.3.2
+
+## 0.1.0 - 24 March 2025
 
 DIVOC (Digital Infrastructure for Verifiable Open Credentialing) is an
 open-source platform that enables countries to digitally orchestrate large-scale