+++ packages/odoo/CHANGELOG.md
@@ -1,13 +1,37 @@
 # @openfn/language-odoo
 
-## 1.0.3
+## 1.0.6 - 22 April 2025
 
 ### Patch Changes
 
-- Updated dependencies [23ccb01]
+- 99e4b48: - Better handling of HTML content in http requests
+  - When logging HTTP requests, include query parameters
+- Updated dependencies \[99e4b48]
+- Updated dependencies \[13bf08f]
+  - @openfn/language-common@2.4.0
+
+## 1.0.5 - 16 April 2025
+
+### Patch Changes
+
+- Updated dependencies \[b089c56]
+  - @openfn/language-common@2.3.3
+
+## 1.0.4 - 11 April 2025
+
+### Patch Changes
+
+- Updated dependencies \[d7105c0]
+  - @openfn/language-common@2.3.2
+
+## 1.0.3 - 14 March 2025
+
+### Patch Changes
+
+- Updated dependencies \[23ccb01]
   - @openfn/language-common@2.3.1
 
-## 1.0.2
+## 1.0.2 - 12 March 2025
 
 ### Patch Changes
 
@@ -22,7 +46,7 @@
   });
   ```
 
-## 1.0.1
+## 1.0.1 - 24 February 2025
 
 ### Patch Changes
 
@@ -32,6 +56,6 @@
 - 2db9f8d: Update the third argument in `read()` examples and documentation to
   explicitly indicate that it accepts an array of strings.
 
-## 1.0.0
+## 1.0.0 - 18 February 2025
 
 Initial release.